requirements.txt
fastapi
uvicorn
jinja2
mysql-connector-python
python-multipart